Cleaning up my code
